Change the mcords load scripts to one file only
Two files is clutter.
